# openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: ScamNemesis API
  description: |
    REST API for the ScamNemesis fraud reporting and tracking platform.

    ## Authentication
    The API supports three authentication methods:
    - **Bearer Token**: JWT access token in Authorization header
    - **HttpOnly Cookie**: Automatic for browser clients
    - **API Key**: X-API-Key header for programmatic access

    ## Rate Limiting
    - Standard users: 100 requests/hour per endpoint
    - Admin users: 1000 requests/hour (10x multiplier)
    - Auth endpoints: 20 requests/15min (brute force protection)

    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional fields.
  version: 1.0.0
  contact:
    name: ScamNemesis Support
    url: https://scamnemesis.com/contact
    email: support@scamnemesis.com
  license:
    name: Proprietary
    url: https://scamnemesis.com/terms
servers:
  - url: https://scamnemesis.com/api/v1
    description: Production server
  - url: https://staging.scamnemesis.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - ApiKeyAuth: []
paths:
...
  /user/api-keys:
    get:
      tags: [User API Keys]
      summary: List user's API keys
      description: |
        Get a paginated list of the authenticated user's API keys.
        Full key values are never returned; only key previews (last 8 characters).
      operationId: listUserApiKeys
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of API keys with statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [User API Keys]
      summary: Create new API key
      description: |
        Create a new API key for the authenticated user.

        **Important**: The full API key is returned only in the response to this
        request. It cannot be retrieved later. Store it securely.

        **Limits**:
        - Maximum 10 active API keys per user
        - Keys can be set to expire between 1-365 days

        **Available Scopes** (users can only select from these):
        - `reports:read`: Read access to fraud reports
        - `search:read`: Access to search functionality
        - `stats:read`: Access to public statistics
      operationId: createUserApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserApiKeyRequest'
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserApiKeyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Maximum API keys limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/api-keys/{id}/rotate:
    post:
      tags: [User API Keys]
      summary: Rotate API key
      description: |
        Generate a new key value for an existing API key.
        The old key immediately stops working.

        **Important**: The new full API key is returned only in this response.
        Store it securely as it cannot be retrieved later.

        This operation resets the request counter and last used timestamp.
      operationId: rotateUserApiKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateUserApiKeyResponse'
        '400':
          description: Key is expired or revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: API key not found

  # ============ EVIDENCE ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly cookie authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Webhook secret for external services
  parameters:
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Request ID for debugging
        timestamp:
          type: string
          format: date-time
      required: [error, message]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            issues:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    UserApiKeyScope:
      type: string
      enum: [reports:read, search:read, stats:read]
      description: |
        Available scopes for user-created API keys:
        - `reports:read`: Read access to fraud reports
        - `search:read`: Access to search functionality
        - `stats:read`: Access to public statistics
    CreateUserApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          pattern: '^[a-zA-Z0-9\s\-_]+$'
          description: Name for the API key (letters, numbers, spaces, hyphens, underscores only)
          example: My Production Key
        description:
          type: string
          maxLength: 500
          description: Optional description
          example: Used for production integration
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/UserApiKeyScope'
          default: [reports:read, search:read, stats:read]
          description: Permission scopes for this key
        expires_in_days:
          type: integer
          minimum: 1
          maximum: 365
          default: 365
          description: Number of days until the key expires
      required: [name]
    CreateUserApiKeyResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique API key ID
        api_key:
          type: string
          description: |
            **IMPORTANT**: This is the only time the full API key will be shown.
            Store it securely. It cannot be retrieved later.
          example: scn_YOUR_API_KEY_HERE
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [personal]
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/UserApiKeyScope'
        rate_limit:
          type: integer
          description: Requests per hour
          example: 1000
        daily_limit:
          type: integer
          nullable: true
          description: Maximum requests per day
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        message:
          type: string
          example: API key created successfully. Store it securely - it won't be shown again.
    UserApiKeySummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [personal]
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/UserApiKeyScope'
        key_preview:
          type: string
          description: Last 8 characters of the key for identification
          example: sk_live_...abcd1234
        is_active:
          type: boolean
        is_expired:
          type: boolean
        days_until_expiry:
          type: integer
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        request_count:
          type: integer
          description: Total number of requests made with this key
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserApiKeyListResponse:
      type: object
      properties:
        api_keys:
          type: array
          items:
            $ref: '#/components/schemas/UserApiKeySummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        stats:
          type: object
          properties:
            active_keys:
              type: integer
              description: Number of active (non-revoked) keys
            total_requests:
              type: integer
              description: Total requests across all keys
            max_keys_allowed:
              type: integer
              example: 10
            keys_remaining:
              type: integer
              description: How many more keys can be created
...
    RotateUserApiKeyResponse:
      type: object
      properties:
        id:
          type: string
        api_key:
          type: string
          description: |
            **IMPORTANT**: This is the only time the new API key will be shown.
            The old key is now invalid. Store this securely.
          example: scn_YOUR_NEW_API_KEY_HERE
        name:
          type: string
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/UserApiKeyScope'
        request_count:
          type: integer
          description: Reset to 0 after rotation
          example: 0
        previous_request_count:
          type: integer
          description: Request count before rotation
        rotated_at:
          type: string
          format: date-time
        message:
          type: string
          example: API key rotated successfully. Store the new key securely - it won't be shown again.

    # ============ PARTNER API SCHEMAS ============
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Authentication required
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...